Rank 7

+++ game.js
@@ -1,10 +1,14 @@
-player={num:0,
+player={timePlayed:0,
+num:0,
+totalNum:0,
 rank:1,
+rankTime:0,
 mults:[1,1,1],
 multBought:[0,0,0],
-version:1}
-time=new Date().getTime()
-rankReqs=[{num:10},{num:150},{num:2000},{num:5e5},{num:1e10}]
+time:0,
+version:1,
+build:2}
+rankReqs=[{num:10},{num:150},{num:2000},{num:5e5},{num:1e10},{num:1e11}]
 costs=[10,2000,5e5]
 tab='main'
 oldTab=tab
@@ -31,15 +35,24 @@ function gameInit() {
 }
 
 function gameTick() {
-	var newTime=new Date().getTime()
-	var diff=(newTime-time)/1000
-	time=newTime
-	
-	player.num+=diff*player.mults[0]*player.mults[1]*player.mults[2]
-	rankUp()
+	if (player.time>0) {
+		var newTime=new Date().getTime()
+		var diff=(newTime-player.time)/1000
+		player.num+=diff*player.mults[0]*player.mults[1]*player.mults[2]
+		player.totalNum+=diff*player.mults[0]*player.mults[1]*player.mults[2]
+		player.timePlayed+=diff
+		player.rankTime+=diff
+		rankUp()
+	}
+	player.time=newTime
 	
 	if (player.rank>2) {
 		showElement('tabs','block')
+		if (player.rank>6) {
+			showElement('tabstatsbutton','inline')
+		} else {
+			hideElement('tabstatsbutton')
+		}
 	} else {
 		hideElement('tabs')
 	}
@@ -73,6 +86,11 @@ function gameTick() {
 		showElement('numberSmall')
 		updateElement('numberSmall',format(player.num))
 	}
+	if (tab=='stats') {
+		updateElement('timePlayed','You have played for '+formatTime(player.timePlayed)+'.')
+		updateElement('totalNumber','You increased the number by '+format(player.totalNum)+' in total.')
+		updateElement('timeRank','You ranked up '+formatTime(player.rankTime)+' ago.')
+	}
 }
 
 function updateElement(elem,id) {
@@ -87,19 +105,41 @@ function hideElement(elem) {
 	document.getElementById(elem).style.display='none'
 }
 
-function format(num,decimalPoints=0,offset=0) {
+function format(num,decimalPoints=2,offset=0) {
 	if (isNaN(num)) {
 		return '?'
 	} else if (num==1/0) {
 		return 'Infinite'
 	} else {
 		var abbid=Math.max(Math.floor(Math.log10(Math.abs(num))/3)-offset,0)
-		var mantissa=Math.round(num/Math.pow(1000,abbid)*Math.pow(10,(abbid>0&&decimalPoints<2)?2:decimalPoints))/Math.pow(10,(abbid>0&&decimalPoints<2)?2:decimalPoints)
+		var mantissa=num/Math.pow(1000,abbid)
+		var log=Math.floor(Math.log10(mantissa))
+		mantissa=(abbid==0)?Math.round(mantissa):Math.round(mantissa*Math.pow(10,3*offset+decimalPoints-log))/Math.pow(10,3*offset+decimalPoints-log)
 		if (mantissa==Math.pow(1000,1+offset)) {
 			mantissa=mantissa/1000
 			abbid+=1
 		}
 		return mantissa+abbs[abbid]
+	}
+}
+
+function formatTime(s) {
+	if (s < 1) {
+		return Math.floor(s*1000)+' milliseconds'
+	} else if (s < 60) {
+		return Math.floor(s*100)/100+' seconds'
+	} else if (s < 3600) {
+		return Math.floor(s/60)+' minutes and '+Math.floor(s%60)+' seconds'
+	} else if (s < 86400) {
+		return Math.floor(s/3600)+' hours, '+Math.floor(s/60%60)+' minutes, and '+Math.floor(s%60)+' seconds'
+	} else if (s < 2629746) {
+		return Math.floor(s/86400)+' days, '+Math.floor(s/3600%24)+' hours, '+Math.floor(s/60%60)+' minutes, and '+Math.floor(s%60)+' seconds'
+	} else if (s < 31556952) {
+		return Math.floor(s/2629746)+' months, '+Math.floor(s%2629746/86400)+' days, '+Math.floor(s%2629746/3600%24)+' hours, '+Math.floor(s%2629746/60%60)+' minutes, and '+Math.floor(s%2629746%60)+' seconds'
+	} else if (s < Infinity) {
+		return format(Math.floor(s/31556952))+' years, '+Math.floor(s/2629746%12)+' months, '+Math.floor(s%2629746/86400)+' days, '+Math.floor(s%2629746/3600%24)+' hours, '+Math.floor(s%2629746/60%60)+' minutes, and '+Math.floor(s%2629746%60)+' seconds'
+	} else {
+		return 'Infinite'
 	}
 }
 
@@ -124,7 +164,7 @@ function rankUp() {
 			if (player[req]<rankReqs[player.rank-1][req]) ableToRankUp=false
 		}
 	}
-	if (ableToRankUp) {player.rank+=1;updateRankText()}
+	if (ableToRankUp) {player.rank+=1;player.rankTime=0;updateRankText()}
 }
 
 function updateCosts() {
@@ -160,7 +200,23 @@ function save() {
 
 function load(savefile) {
 	try {
-		player=JSON.parse(atob(savefile))
+		savefile=JSON.parse(atob(savefile))
+		if (savefile.build==undefined) {
+			savefile.version=1
+			savefile.build=1
+		}
+		if (savefile.version<2) {
+			if (savefile.build<2) {
+				savefile.time=new Date().getTime()
+				savefile.timePlayed=0
+				savefile,totalNum=0
+				savefile.rankTime=0
+			}
+		}
+		savefile.version=player.version
+		savefile.build=player.build
+		
+		player=savefile
 		updateRankText()
 		updateCosts()
 		console.log('Game loaded!')
@@ -185,13 +241,15 @@ function importSave() {
 }
 
 function reset() {
-	player.num=0
-	player.rank=1
-	player.mults=[1,1,1]
-	player.multBought=[0,0,0]
-	localStorage.clear('saveRanks')
-	
-	updateRankText()
-	updateCosts()
-	tab='main'
+	if (confirm('Are you sure to reset your save? You can\'t undo your action!')) {
+		player.num=0
+		player.rank=1
+		player.mults=[1,1,1]
+		player.multBought=[0,0,0]
+		localStorage.clear('saveRanks')
+		
+		updateRankText()
+		updateCosts()
+		tab='main'
+	}
 }
